Fix GObject Introspection paths to work with 18.08
Freedesktop-SDK updates from GNU pkg-config to the newer pkgconf,
however pkgconf does not allow environmental variables to override
pkg-config variables. Thus libgee and granite both try to install
to girdirs and typelibdirs in /usr rather than /app, which are
read only and crash the build. This patch provides workarounds
for both, although the patch for granite is a little hacky.